+++ src/ueb_03/TCP/META-INF/MANIFEST.MF
@@ -1,0 +1,3 @@
+Manifest-Version: 1.0
+Main-Class: oxoo2a.Main
+
