+++ src/ueb_03/UDP/META-INF/MANIFEST.MF
@@ -1,0 +1,2 @@
+Manifest-Version: 1.0
+Main-Class: ueb_03.UDP.Main
